refactor(cli): share the tolerant argv scan and export the gateway defaults
