+++ packages/server/src/serve.ts
@@ -2,14 +2,11 @@ import type { ToolDefinition } from "@earendil-works/pi-coding-agent";
 
 import { AttachmentStore } from "#core/attachments/AttachmentStore";
 import { SessionRegistry } from "#core/session/SessionRegistry";
+import { Cli as Argv } from "#core/shared/Cli";
 import { Tools } from "#core/shared/Tools";
 import { defaultAttachmentsRoot } from "./AttachmentEndpoint";
 import { SendFileTool } from "./SendFileTool";
-import { WsGateway } from "./WsGateway";
-
-const DEFAULT_PORT = "4319";
-/** Loopback only: the server has full host access and no authentication. */
-const DEFAULT_HOSTNAME = "127.0.0.1";
+import { DEFAULT_HOSTNAME, DEFAULT_PORT, WsGateway } from "./WsGateway";
 
 export type Cli = {
   readonly port: string;
@@ -18,34 +15,15 @@ export type Cli = {
   readonly clientDir: string | undefined;
 };
 
-/**
- * Tolerant on purpose: the same argv reaches here through `pim --mode web`,
- * so unknown flags and the mode positional must not be fatal.
- */
 export function parseArgs(args: ReadonlyArray<string>): Cli {
   // PORT is honoured because a supervisor or container assigns it, but an
   // explicit `--port` always wins.
-  let port = process.env["PORT"] ?? DEFAULT_PORT;
+  let port = process.env["PORT"] ?? String(DEFAULT_PORT);
   let hostname = DEFAULT_HOSTNAME;
   let cwd = process.cwd();
   let clientDir: string | undefined;
 
-  for (let i = 0; i < args.length; i++) {
-    const arg = args[i]!;
-    if (!arg.startsWith("--")) {
-      continue;
-    }
-    const eqIdx = arg.indexOf("=");
-    const key = eqIdx >= 0 ? arg.slice(0, eqIdx) : arg;
-    const inline = eqIdx >= 0 ? arg.slice(eqIdx + 1) : undefined;
-    const take = (): string | undefined => {
-      if (inline !== undefined) {
-        return inline;
-      }
-      i += 1;
-      return args[i];
-    };
-
+  Argv.scan(args, (key, take) => {
     switch (key) {
       case "--port":
         port = take() ?? port;
@@ -62,7 +40,7 @@ export function parseArgs(args: ReadonlyArray<string>): Cli {
       default:
         break;
     }
-  }
+  });
   return { port, hostname, cwd, clientDir };
 }
